+++ packages/react-tinacms/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.10](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.10-alpha.1...react-tinacms@0.9.10) (2020-02-18)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
+## [0.9.10-alpha.1](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.10-alpha.0...react-tinacms@0.9.10-alpha.1) (2020-02-16)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
+## [0.9.10-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.9...react-tinacms@0.9.10-alpha.0) (2020-02-14)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
 ## [0.9.9](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.9-alpha.1...react-tinacms@0.9.9) (2020-02-11)
 
 **Note:** Version bump only for package react-tinacms